Bib Reference Seems to Work
I am not sure why it is working not but it is. I think that it all depends on the order in which you compile the document. 

In Texmaker I use quick build but I used it after I complied in the following order: PDFLatex >> Bibtex >> View Pdf. 

I am not sure if this had anything to do with it but it seems to work. 

P.S. This is tested in chapter 1